% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{compute_Gamma_from_rho}
\alias{compute_Gamma_from_rho}
\title{Given the maximal probability of treatment within a matched pair in
a sensitivity analysis, compute the associated sensitivity parameter
Gamma.  Not meant to be called directly by users.}
\usage{
compute_Gamma_from_rho(root_equation)
}
\arguments{
\item{root_equation}{A function of treatment probability rho that
achieves value zero for a situation of interest.}
}
\value{
The value of Gamma (a single numeric value) associated with
    the value of rho that solves the input root equation.
}
\description{
Given the maximal probability of treatment within a matched pair in
a sensitivity analysis, compute the associated sensitivity parameter
Gamma.  Not meant to be called directly by users.
}
