% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_Gamma.R
\name{fixed_max_Gamma}
\alias{fixed_max_Gamma}
\title{Report the largest sensitivity parameter Gamma for which the
sensitivity analysis for a uniform general signed rank test rejects
the null hypothesis at level alpha., i.e. the largest value of
Gamma for which the upper uniform boundary intersects the random
walk induced by the observed data.}
\usage{
fixed_max_Gamma(outcomes, score_fn = "sign", alpha = 0.05)
}
\arguments{
\item{outcomes}{A vector of matched pair differences.}

\item{score_fn}{A string indicating which score function should be
used.  Must be one of 'sign' (sign score function), 'wsrt'
(Wilcoxon signed rank score function), 'normal_scores'
(normal score function) and 'redescending'.}

\item{alpha}{Rate of type I error control for the hypothesis test.}
}
\value{
The maximum sensitivity parameter value (a single numeric
    value no smaller than 1).
}
\description{
Report the largest sensitivity parameter Gamma for which the
sensitivity analysis for a uniform general signed rank test rejects
the null hypothesis at level alpha., i.e. the largest value of
Gamma for which the upper uniform boundary intersects the random
walk induced by the observed data.
}
\examples{
data(mercury2)
fixed_max_Gamma(mercury2, 'sign')
uniform_max_Gamma(mercury2, 'sign')
}
\references{
Howard, S.R., and Pimentel, S.D. (2020), "The uniform general
signed rank test and its design sensitivity." arXiv:1904.08895.
}
