% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{fixed_pval}
\alias{fixed_pval}
\title{Report the p-value from a general signed rank test, or from an
    associated sensitivity analysis.}
\usage{
fixed_pval(outcomes, score_fn = "sign", Gamma = 1, tol = 1e-15)
}
\arguments{
\item{outcomes}{A vector of matched pair differences.}

\item{score_fn}{A string indicating which score function should be
used.  Must be one of 'sign' (sign score function), 'wsrt'
(Wilcoxon signed rank score function), 'normal_scores'
(normal score function) and 'redescending'.}

\item{Gamma}{Sensitivity parameter (>=1) describing the degree
of unmeasured confounding present.  The default case Gamma = 1
corresponds to no unmeasured confounding.}

\item{tol}{A minimum tolerance for distinguishing p-values from 0
and 1.  P-values found to be below \code{tol} or above 1 -
\code{tol} will be reported as 0 and 1 respectively.}
}
\value{
A p-value.
}
\description{
Report the p-value from a general signed rank test, or from an
    associated sensitivity analysis.
}
\examples{
data(mercury2)
uniform_pval(mercury2, 'sign') #test assuming no unobserved confounding
uniform_pval(mercury2, 'sign', Gamma = 5) #with large amount of confounding
fixed_pval(mercury2, 'sign', Gamma = 5) #fixed test is more sensitive
}
\references{
Howard, S.R., and Pimentel, S.D. (2020), "The uniform general
signed rank test and its design sensitivity." arXiv:1904.08895.
}
