% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_scores}
\alias{get_scores}
\title{Given a score function and a sample size n, returns scores for points
spaced equally at 1/(n+1) intervals within the unit interval.  Not meant
to be called directly by users.}
\usage{
get_scores(score_fn, num_units)
}
\arguments{
\item{score_fn}{A score function taking a numeric vector of elements in
$[0,1]$ and returning another nonnegative numeric vector of the same
length.}

\item{num_units}{The number of matched pairs in the study, an integer.}
}
\value{
A vector of length \code{num_units} given the scores for
    equispaced points within the unit interval.
}
\description{
Given a score function and a sample size n, returns scores for points
spaced equally at 1/(n+1) intervals within the unit interval.  Not meant
to be called directly by users.
}
