% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_scores_and_T_by_score}
\alias{get_scores_and_T_by_score}
\title{Generate the score for each pair difference and the observed
value of the random walk T using the raw pair differences and
the score function. Not meant to be called directly by users.}
\usage{
get_scores_and_T_by_score(outcomes, score_fn)
}
\arguments{
\item{outcomes}{A vector of matched pair differences.}

\item{score_fn}{A string indicating which score function should be
used.  Must be one of 'sign' (sign score function), 'wsrt'
(Wilcoxon signed rank score function), 'normal_scores'
(normal score function) and 'redescending'.}
}
\value{
A list containing two elements, \code{scores} and
    \code{T}. Vector \code{scores} gives the score for each
    pair difference in non-decreasing order. Vector \code{T}
    gives the corresponding random walk for the observed data.
}
\description{
Generate the score for each pair difference and the observed
value of the random walk T using the raw pair differences and
the score function. Not meant to be called directly by users.
}
