% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_u_statistic_score_fn}
\alias{get_u_statistic_score_fn}
\title{The approximate score associated with the u-statistic of Rosenbaum (2011),
which counts, for all subsets of size \code{m} of the set of matched pairs,
the number of positive differences in the pairs with the \code{m_lower} to
\code{m_upper} smallest absolute differences, inclusive. Not meant to be
called directly by users.}
\usage{
get_u_statistic_score_fn(m, m_lower, m_upper)
}
\arguments{
\item{m}{Size of subsets to consider.}

\item{m_lower}{Lower boundary in the within-subset rank-ordering for pairs to
consider.  Must be no larger than \code{m}.}

\item{m_upper}{Upper boundary in the within-subset rank-ordering for pairs to
consider.  Must be no larger than \code{m} and no smaller than
$\code{m_lower}.}
}
\value{
A function mapping numeric values in $[0,1]$ to nonnegative numeric
    values.
}
\description{
The approximate score associated with the u-statistic of Rosenbaum (2011),
which counts, for all subsets of size \code{m} of the set of matched pairs,
the number of positive differences in the pairs with the \code{m_lower} to
\code{m_upper} smallest absolute differences, inclusive. Not meant to be
called directly by users.
}
\references{
Rosenbaum, P. R. (2011). A new u-statistic with superior design
    sensitivity in matched observational studies. \emph{Biometrics}, 67(3),
    1017-1027.
}
