% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mercury2}
\alias{mercury2}
\title{Mercury Concentration Differences for Matched NHANES Respondents.}
\format{A vector with 1,672 elements, each one a pair
difference.}
\source{
\url{https://arxiv.org/pdf/1904.08895.pdf}
}
\usage{
mercury2
}
\description{
Differences in total mercury concentration in the blood (in
micrograms per mole) for subjects in 1,672 matched pairs.
Each pair contains two respondents to the National Health and
Nutrition Examination Survey (NHANES) during the period 2007-
2016, one who reported consuming 15 or more servings of fish
per month, and one who reported consuming two or fewer
servings per month.  Positive differences indicate that the
respondent consuming more fish had a higher mercury
concentration.  These data are described in greater detail
in:
}
\details{
Howard, S.R., and Pimentel, S.D. (2020), "The uniform general
signed rank test and its design sensitivity." arXiv:1904.08895.
}
\keyword{datasets}
