% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{score_fns}
\alias{score_fns}
\title{Given the name of a score function as input, return the score
function itself. Not meant to be called directly by users.}
\usage{
score_fns(scorename)
}
\arguments{
\item{scorename}{A string indicating which score function should be
used.  Must be one of 'sign' (sign score function), 'wsrt'
(Wilcoxon signed rank score function), 'normal_scores'
(normal score function) and 'redescending'.}
}
\value{
A score function taking a numeric vector of elements in $[0,1]$
    and returning another nonnegative numeric vector of the same length.
}
\description{
Given the name of a score function as input, return the score
function itself. Not meant to be called directly by users.
}
